Improved the general code slightly further, and played with the model structure in an attempt improve its training ability.
Also included a scheduler that decreases the learning rate a little every 500 steps of the training.

Finally, added saving and loading of the models and their learning rate, in addition to the criterion's, optimizer's and scheduler's state dictionaries so that they can all be saved and reloaded at any point, and continued to be trained with the same values easily.